removed contact selection and renamed customer selection to contact selection
