+++ src/Sulu/Bundle/ContactBundle/DependencyInjection/Configuration.php
@@ -43,16 +43,6 @@ class Configuration implements ConfigurationInterface
                                 ->end()
                             ->end()
                         ->end()
-                        ->arrayNode('customer')
-                            ->addDefaultsIfNotSet()
-                            ->children()
-                                ->scalarNode('template')
-                                    ->defaultValue(
-                                        'SuluContactBundle:Template:content-types/customer-selection.html.twig'
-                                    )
-                                ->end()
-                            ->end()
-                        ->end()
                     ->end()
                 ->end()
                 ->arrayNode('defaults')
