--- src/Sulu/Bundle/ContactBundle/Resources/public/js/validation/types/customer-selection.js
@@ -1,50 +1,0 @@
-/*
- * This file is part of the Husky Validation.
- *
- * (c) MASSIVE ART WebServices GmbH
- *
- * This source file is subject to the MIT license that is bundled
- * with this source code in the file LICENSE.
- *
- */
-
-define(['type/default'], function(Default) {
-
-    'use strict';
-
-    var dataChangedHandler = function (data, $el) {
-        App.emit('sulu.preview.update', $el, data);
-        App.emit('sulu.content.changed');
-    };
-
-    return function($el, options) {
-        var defaults = {},
-
-            subType = {
-                initializeSub: function() {
-                    var dataChangedEvent = 'sulu.customer-selection.' + options.instanceName + '.data-changed';
-
-                    App.off(dataChangedEvent, dataChangedHandler);
-                    App.on(dataChangedEvent, dataChangedHandler);
-                },
-
-                setValue: function(value) {
-                    App.dom.data($el, 'customer-selection', value);
-                },
-
-                getValue: function() {
-                    return App.dom.data($el, 'customer-selection');
-                },
-
-                needsValidation: function() {
-                    return false;
-                },
-
-                validate: function() {
-                    return true;
-                }
-            };
-
-        return new Default($el, defaults, options, 'customerSelection', subType);
-    };
-});
