+++ src/Sulu/Bundle/ContactBundle/Resources/translations/sulu/backend.de.xlf
@@ -624,20 +624,12 @@
                 <source>contact.accounts.logo-dropzone-text</source>
                 <target>Bild ablegen oder Klicken um das Logo zu ändern</target>
             </trans-unit>
-            <trans-unit id="235b14f26g712345e44gs2e9335a8d50" resname="contact-selection.no-contact-selected">
-                <source>contact-selection.no-contact-selected</source>
-                <target>Keine Person ausgewählt</target>
-            </trans-unit>
-            <trans-unit id="235b14f26g712345e44gs2e9335a8d51" resname="contact-selection.add">
-                <source>contact-selection.add</source>
-                <target>Personen auswählen</target>
-            </trans-unit>
-            <trans-unit id="235b14f26g712345e44gs2e9335a8d52" resname="customer-selection.no-customer-selected">
-                <source>customer-selection.no-customer-selected</source>
+            <trans-unit id="235b14f26g712345e44gs2e9335a8d52" resname="contact-selection.no-customer-selected">
+                <source>customer-selection.no-contact-selected</source>
                 <target>Keine Person oder Organisation ausgewählt</target>
             </trans-unit>
-            <trans-unit id="235b14f26g712345e44gs2e9335a8d53" resname="customer-selection.add">
-                <source>customer-selection.add</source>
+            <trans-unit id="235b14f26g712345e44gs2e9335a8d53" resname="contact-selection.add">
+                <source>contact-selection.add</source>
                 <target>Personen oder Organisation auswählen</target>
             </trans-unit>
         </body>
