+++ src/Sulu/Bundle/ContactBundle/Resources/translations/sulu/backend.en.xlf
@@ -630,13 +630,13 @@
                 <source>contact-selection.add</source>
                 <target>Choose person</target>
             </trans-unit>
-            <trans-unit id="235b14f26g712345e44gs2e9335a8d52" resname="customer-selection.no-customer-selected">
-                <source>customer-selection.no-customer-selected</source>
-                <target>No person or organization selected</target>
+            <trans-unit id="235b14f26g712345e44gs2e9335a8d52" resname="contact-selection.no-customer-selected">
+                <source>customer-selection.no-contact-selected</source>
+                <target>Keine Person oder Organisation ausgewählt</target>
             </trans-unit>
-            <trans-unit id="235b14f26g712345e44gs2e9335a8d53" resname="customer-selection.add">
-                <source>customer-selection.add</source>
-                <target>Choose person or organization</target>
+            <trans-unit id="235b14f26g712345e44gs2e9335a8d53" resname="contact-selection.add">
+                <source>contact-selection.add</source>
+                <target>Personen oder Organisation auswählen</target>
             </trans-unit>
         </body>
     </file>
