+++ src/Sulu/Bundle/ContactBundle/Resources/views/Template/content-types/contact-selection.html.twig
@@ -7,5 +7,7 @@
      data-type-instance-name="{{ id|raw }}"
      data-aura-component="contact-selection@sulucontact"
      data-aura-instance-name="{{ id|raw }}"
-     data-aura-url="/admin/api/contacts?flat=true">
+     data-aura-url="/admin/api/customers"
+     data-aura-contact-url="/admin/api/contacts?flat=true"
+     data-aura-account-url="/admin/api/accounts?flat=true">
 </div>
