--- src/Sulu/Bundle/ContactBundle/Resources/views/Template/content-types/customer-selection.html.twig
@@ -1,13 +1,0 @@
-<div id="{{ id|raw }}"
-     class="hide-in-sortmode"
-     data-property='{{ property|json_encode }}'
-     data-mapper-property="{{ property.name }}"
-     data-form="true"
-     data-type="customer-selection"
-     data-type-instance-name="{{ id|raw }}"
-     data-aura-component="customer-selection@sulucontact"
-     data-aura-instance-name="{{ id|raw }}"
-     data-aura-url="/admin/api/customers"
-     data-aura-contact-url="/admin/api/contacts?flat=true"
-     data-aura-account-url="/admin/api/accounts?flat=true">
-</div>
